Iterable refactoring
- Iterable is now for synchronous and deterministic iterators

- IterableIO is for synchronous and non-deterministic iterators (async and any other iterators where the outcome may be different each time you run it).

- Also started the process of making collection Count into a long to support large collections.

+++ LanguageExt.Core/Deriving/Foldable.cs
@@ -254,7 +254,7 @@ public static partial class Deriving
         /// <summary>
         /// Find the element at the specified index or `None` if out of range
         /// </summary>
-        static Option<A> Foldable<Supertype>.At<A>(int index, K<Supertype, A> ta) =>
+        static Option<A> Foldable<Supertype>.At<A>(long index, K<Supertype, A> ta) =>
             Subtype.At(index, Supertype.Transform(ta));
 
         /// <summary>
