+++ LanguageExt.Core/Traits/Foldable/Foldable.Module.cs
@@ -175,7 +175,7 @@ public static class Foldable
     /// than via element-by-element counting, should provide a specialised
     /// implementation.
     /// </summary>
-    public static int count<T, A>(K<T, A> ta) 
+    public static long count<T, A>(K<T, A> ta) 
         where T : Foldable<T> =>
         ta.Count;
 
@@ -260,7 +260,7 @@ public static class Foldable
     /// actions from left to right, and ignore the results.  For a version that
     /// doesn't ignore the results see `Traversable.traverse`.
     /// </summary>
-    public static Unit iter<T, A>(Action<int, A> f, K<T, A> ta) 
+    public static Unit iter<T, A>(Action<long, A> f, K<T, A> ta) 
         where T : Foldable<T> =>
         ta.Iter(f);
